feat #57: MenuItem 컴포넌트 분리 및 간단한 props 설정
