+++ src/components/dropdown/MenuItem.tsx
@@ -1,0 +1,40 @@
+import { ReactNode } from 'react';
+import { styled } from 'styled-components';
+
+export function MenuItem({
+  onClick,
+  children,
+}: {
+  onClick: () => void;
+  children: ReactNode;
+}) {
+  return <Item onClick={onClick}>{children}</Item>;
+}
+
+const Item = styled.li`
+  width: 240px;
+  display: flex;
+  padding: 8px;
+  align-items: center;
+  gap: 8px;
+  align-self: stretch;
+  padding: 16px;
+  border-bottom: 1px solid ${({ theme }) => theme.color.neutralBorder};
+  font: ${({ theme }) => theme.font.availableDefault16};
+  color: ${({ theme }) => theme.color.neutralTextStrong};
+
+  &:first-child {
+    border-top-left-radius: 12px;
+    border-top-right-radius: 12px;
+  }
+
+  &:last-child {
+    border-bottom-left-radius: 12px;
+    border-bottom-right-radius: 12px;
+    border-bottom: none;
+  }
+
+  &:hover {
+    background-color: ${({ theme }) => theme.color.neutralBackgroundBlur};
+  }
+`;
